Append events on fan-in instead of clearing prior sources
